Deep arquitecture and folder structure change
